// src/app/interfaces/speciality/get-specialities-api-response.interface.ts
import { IUserCreator } from "../user/user-creator.interface";

export interface IGetSpecialitiesApiResponse {
    id: string;
    code: string;
    name: string;
    description?: string;
    status: boolean;
    created_at: string;
    updated_at: string;
    deleted_at?: string;
    user_creator: IUserCreator;
}